# backend/docs/swagger.yaml
basePath: /api
definitions:
  api.NetworkLogRequest:
    properties:
      host_id:
        type: string
      log_data:
        type: string
      log_type:
        type: string
    required:
    - host_id
    - log_data
    - log_type
    type: object
...
  dto.WindowsArtifactDTO:
    properties:
      hash:
        type: string
      host_id:
        type: string
      id:
        type: integer
      metadata:
        additionalProperties: true
        type: object
      owner:
        type: string
      path:
        type: string
      permissions:
        type: string
      size:
        type: integer
      threat_level:
        type: integer
      timestamp:
        type: string
      type:
        type: string
      value:
        type: string
    type: object
...
  models.NetworkTraffic:
    type: object
...
  models.TrafficStats:
    properties:
      average_duration:
        type: number
      total_bytes_recv:
        type: integer
      total_bytes_sent:
        type: integer
      total_packets_recv:
        type: integer
      total_packets_sent:
        type: integer
    type: object
  models.WindowsArtifact:
    type: object
...
  requests.IsolateHostRequest:
    properties:
      duration:
        description: в минутах, 0 = бессрочно
        type: integer
      host_id:
        type: string
      reason:
        type: string
    required:
    - host_id
    - reason
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Railgun AI-driven SIEM API
  title: Railgun API
  version: "1.0"
paths:
...
  /artifacts:
    post:
      consumes:
      - application/json
      description: Создает новую запись артефакта Windows в системе
      parameters:
      - description: Данные артефакта
        in: body
        name: artifact
        required: true
        schema:
          $ref: '#/definitions/models.WindowsArtifact'
      produces:
      - application/json
      responses:
        "201":
          description: Успешное сохранение и ID записи
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Сохранить артефакт
      tags:
      - Artifacts
  /artifacts/host/{hostId}:
    get:
      description: Возвращает список артефактов для конкретного хоста с поддержкой
        пагинации
      parameters:
      - description: ID хоста
        in: path
        name: hostId
        required: true
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Количество на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Объект с массивом data и мета-данными
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить артефакты хоста
      tags:
      - Artifacts
  /artifacts/id/{id}:
    get:
      description: Возвращает детальную информацию об одном артефакте
      parameters:
      - description: ID артефакта
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.WindowsArtifactDTO'
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Артефакт не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получить артефакт по ID
      tags:
      - Artifacts
  /artifacts/search:
    get:
      description: Поиск по артефактам с фильтрацией по типу, критичности и текстовому
        запросу
      parameters:
      - description: Поисковый запрос
        in: query
        name: q
        type: string
      - description: Тип артефакта
        in: query
        name: type
        type: string
      - description: Уровень серьезности
        in: query
        name: severity
        type: string
      - default: 1
        description: Номер страницы
        in: query
        name: page
        type: integer
      - default: 20
        description: Количество на странице
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Результаты поиска и мета-данные
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Поиск артефактов
      tags:
      - Artifacts
...
  /ingest/isolate:
    post:
      consumes:
      - application/json
      description: Блокирует сетевую активность хоста по его ID на указанное время
      parameters:
      - description: Параметры изоляции (duration в минутах)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.IsolateHostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Подтверждение изоляции
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изоляция хоста
      tags:
      - Response
  /ingest/logs:
    post:
      consumes:
      - application/json
      description: Принимает неструктурированные логи (syslog и др.), парсит их и
        преобразует в записи трафика
      parameters:
      - description: Сырые логи и метаданные
        in: body
        name: log_request
        required: true
        schema:
          $ref: '#/definitions/api.NetworkLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Количество обработанных записей (entries_processed)
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Прием сырых логов
      tags:
      - Ingestion
  /ingest/stats/{hostId}:
    get:
      description: Возвращает агрегированную статистику (объемы, протоколы) для конкретного
        хоста за период
      parameters:
      - description: ID хоста
        in: path
        name: hostId
        required: true
        type: string
      - description: Начало (RFC3339)
        in: query
        name: from
        type: string
      - description: Конец (RFC3339)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.TrafficStats'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Статистика трафика хоста
      tags:
      - Analytics
  /ingest/traffic:
    post:
      consumes:
      - application/json
      description: Принимает JSON с сетевой активностью от агентов, сохраняет в БД
        и отправляет в корреляционный движок
      parameters:
      - description: Данные сетевого трафика
        in: body
        name: traffic
        required: true
        schema:
          $ref: '#/definitions/models.NetworkTraffic'
      produces:
      - application/json
      responses:
        "201":
          description: 'status: accepted'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Прием структурированного трафика
      tags:
      - Ingestion
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"